Revert "add(core): Support for Prefix Caching (#85)"
This reverts commit 0794b794dc4a3430554388fbbaf3325c82d68592.
