+++ vllm_rbln/model_executor/models/optimum/decoder_only.py
@@ -14,8 +14,7 @@
 import torch
 import vllm.envs as envs
 from vllm.config import VllmConfig
-
-from vllm_rbln.logger import init_logger
+from vllm.logger import init_logger
 
 from .base import ModelInputForRBLN, version_error
 from .model_base import RBLNOptimumDecoderMixin, RBLNOptimumModelBase
@@ -60,11 +59,7 @@ class RBLNOptimumForCausalLM(RBLNOptimumModelBase, RBLNOptimumDecoderMixin):
         if is_prompt:
             if self.model.prefill_decoder is None:
                 raise version_error
-            if model_input.cached_block_tables:
-                self._copy_cached_kv_blocks(self.model.prefill_decoder,
-                                            model_input.cached_block_tables,
-                                            model_input.cached_lengths,
-                                            block_tables)
+
             return self.model.prefill_decoder(**kwargs).logits
         else:
             self.model.decoder = self.model.decoders[padded_batch_size]
